pyproject.toml
[project]
name = "solrizer"
version = "1.0.0"
dependencies = [
    "beautifulsoup4",
    "edtf",
    "flask",
    "jq",
    "langcodes",
    "plastron-client>=4.4.0",
    "plastron-models>=4.4.0",
    "plastron-rdf>=4.4.0",
    "plastron-repo>=4.4.0",
    "plastron-utils>=4.4.0",
    "pyyaml",
    "python-dotenv",
    "python-iso639",
    "python-jsonpath",
    "pyyaml",
    # require rdflib 7.0.0+ in order to have access
    # to the NamespaceManager.curie() method
    "rdflib>=7.0.0",
    "requests-jwtauth",
    "uritemplate",
    "waitress",
]
requires-python = ">=3.12"

[project.optional-dependencies]
test = [
    "httpretty",
    "pytest",
    "pytest-cov",
    "pytest-datadir",
    "pycodestyle",
    "ruff",
]
dev = [
    "pdoc",
]

[project.scripts]
solrizer = "solrizer.server:run"

[project.entry-points.solrizer_indexers]
content_model = "solrizer.indexers.content_model:content_model_fields"
discoverability = "solrizer.indexers.discoverability:discoverability_fields"
page_sequence = "solrizer.indexers.page_sequence:page_sequence_fields"
iiif_links = "solrizer.indexers.iiif_links:iiif_links_fields"
dates = "solrizer.indexers.dates:date_fields"
facets = "solrizer.indexers.facets:facet_fields"
extracted_text = "solrizer.indexers.extracted_text:extracted_text_fields"
root = "solrizer.indexers.root:root_field"
handles = "solrizer.indexers.handles:handle_fields"
aggregate_fields = "solrizer.indexers.aggregate_fields:aggregate_fields"

[project.entry-points.solrizer_faceters]
admin_set = "solrizer.faceters:AdminSetFacet"
archival_collection = "solrizer.faceters:ArchivalCollectionFacet"
censorship = "solrizer.faceters:CensorshipFacet"
contributor = "solrizer.faceters:ContributorFacet"
creator = "solrizer.faceters:CreatorFacet"
has_ocr = "solrizer.faceters:OCRFacet"
language = "solrizer.faceters:LanguageFacet"
location = "solrizer.faceters:LocationFacet"
presentation_set = "solrizer.faceters:PresentationSetFacet"
publication_status = "solrizer.faceters:PublicationStatusFacet"
publisher = "solrizer.faceters:PublisherFacet"
rdf_type = "solrizer.faceters:RDFTypeFacet"
resource_type = "solrizer.faceters:ResourceTypeFacet"
rights = "solrizer.faceters:RightsFacet"
subject = "solrizer.faceters:SubjectFacet"
visibility = "solrizer.faceters:VisibilityFacet"

[tool.pytest.ini_options]
# these warnings are coming from third-party code
filterwarnings = [
    'ignore:datetime.datetime.utcnow\(\) is deprecated:DeprecationWarning',
    'ignore:ConjunctiveGraph is deprecated:DeprecationWarning',
    "ignore:The 'strip_cdata' option:DeprecationWarning",
    "ignore:distutils Version classes are deprecated:DeprecationWarning",
]
addopts = "--doctest-modules"

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
# preview is required to enable many pycodestyle rules see
# https://docs.astral.sh/ruff/rules/
preview = true

# Enable pycodestyle errors and warnings
select = ["E", "W"]
